backend/requirements.txt
pandas
numpy
matplotlib
pymongo
python-dotenv
